feat(modeling): rename expansions to offsets
