+++ packages/modeling/src/index.d.ts
@@ -8,10 +8,10 @@ export * as text from './text'
 export * as utils from './utils'
 
 export * as booleans from './operations/booleans'
-export * as expansions from './operations/expansions'
 export * as extrusions from './operations/extrusions'
 export * as hulls from './operations/hulls'
 export * as modifiers from './operations/modifiers'
+export * as offsets from './operations/offsets'
 export * as transforms from './operations/transforms'
 
 export as namespace modeling
