+++ packages/modeling/src/index.js
@@ -8,8 +8,8 @@ export * as text from './text/index.js'
 export * as utils from './utils/index.js'
 
 export * as booleans from './operations/booleans/index.js'
-export * as expansions from './operations/expansions/index.js'
 export * as extrusions from './operations/extrusions/index.js'
 export * as hulls from './operations/hulls/index.js'
 export * as modifiers from './operations/modifiers/index.js'
+export * as offsets from './operations/offsets/index.js'
 export * as transforms from './operations/transforms/index.js'
