+++ packages/modeling/src/operations/offsets/extrudePolygon.js
