+++ packages/modeling/src/operations/offsets/index.d.ts
@@ -1,4 +1,4 @@
 export { offset, OffsetOptions } from './offset'
 export { offsetFromPoints } from './offset'
 
-export as namespace expansions
+export as namespace offsets
