+++ packages/modeling/src/operations/offsets/index.js
@@ -1,0 +1,10 @@
+/**
+ * Polygon offsetting is a geometric operation that expands or contracts
+ * shapes by a fixed distance.
+ * @module modeling/offsets
+ * @example
+ * import { offsets } from '@jscad/modeling'
+ * const { offset } = offsets
+ */
+export { offset } from './offset.js'
+export { offsetFromPoints } from './offsetFromPoints.js'
