+++ packages/modeling/src/operations/offsets/offset.d.ts
