+++ packages/modeling/src/operations/offsets/offsetFromPoints.d.ts
