+++ packages/modeling/src/operations/offsets/offsetGeom2.js
