+++ packages/modeling/src/operations/offsets/offsetGeom3.js
