+++ packages/modeling/src/operations/offsets/offsetGeom3.test.js
