+++ packages/modeling/src/operations/offsets/offsetPath2.js
