+++ packages/modeling/src/operations/offsets/offsetShell.js
